[Refactor] - 패키지 위치 정리
